+++ common/src/main/java/com/sonicether/soundphysics/mixin/LocalPlayerMixin.java
@@ -1,0 +1,33 @@
+package com.sonicether.soundphysics.mixin;
+
+import com.sonicether.soundphysics.SoundPhysicsMod;
+import com.sonicether.soundphysics.utils.SoundUtils;
+import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.sounds.SoundEvent;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
+
+@Mixin(LocalPlayer.class)
+public abstract class LocalPlayerMixin extends Entity {
+
+    public LocalPlayerMixin(EntityType<?> entityType, Level level) {
+        super(entityType, level);
+    }
+
+    @ModifyArg(method = "playSound", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/Level;playLocalSound(DDDLnet/minecraft/sounds/SoundEvent;Lnet/minecraft/sounds/SoundSource;FFZ)V"), index = 1)
+    private double playSound(double x, double y, double z, SoundEvent sound, SoundSource category, float volume, float pitch, boolean b) {
+        if (sound == null) {
+            return y;
+        }
+        if (!SoundPhysicsMod.CONFIG.enabled.get()) {
+            return y;
+        }
+        return y + SoundUtils.calculateEntitySoundYOffset(getEyeHeight(), sound);
+    }
+
+}